Move Playbook::Props into its own namespace
