+++ app/pb_kits/playbook/props/base.rb
@@ -1,0 +1,18 @@
+# frozen_string_literal: true
+
+module Playbook
+  module Props
+    class Base
+      def initialize(default: nil)
+        @default = default
+      end
+
+      def value(value)
+        value || @default
+      end
+
+      def validate!(_value)
+      end
+    end
+  end
+end
